ArC: fix hash code of TypeAndQualifiers
The `TypeAndQualifiers` compound is used as a key in hash maps, which
means its equality and hash code must be consistent. Unfortunately,
the implementation of `equals()` is completely custom, because it
needs to ignore `AnnotationInstance.target`, yet the `hashCode()`
implementation blindly delegates to `AnnotationInstance`. That used
to work, because `AnnotationInstance.hashCode()` ignored the `target`
before, but that has never been a stated contract. A completely custom
implementation of equality should always be accompanied by a custom
implementation of hash code. That's what this commit does.