Ignore the module-info from the application in dev mode.
As described in https://github.com/quarkusio/quarkus/issues/17747, the modules are not found in dev mode, breaking the hot reload.

The implemented behavior can be disabled using the `quarkus.live-reload.ignore-module-info` property (in the project descriptor, not in the `application.properties`). Also, a log message will explain the behavior and indicate how to disable it.
The log message is displayed only once.

Fix https://github.com/quarkusio/quarkus/issues/17747 - as described in the issue