Update support information for extension descriptors
In support of quarkusio#8021, process the supported-by attribute to find a supporter name, then create or update an attribute with the supporter id, name with the status(es) for the extension supporter.  If there are existing tags ending with `-support`, a support entry will be created / updated for the supporter(s).  If any
of the support entries are missing a status, the default value of `stable` will be used.

Signed-off-by:Nathan Erwin <nathan.d.erwin@gmail.com>