Infinispan Client: add runtime SPI and deployment SPI modules
Some constants from the Infinispan Client runtime module are moved
to the runtime SPI module, so that they can be used from other modules
that cannot directly depend on the Infinispan Client extension.

Two build items, `InfinispanClientBuildItem` and `InfinispanClientNameBuildItem`,
are moved from the Infinispan Client deployment module to the deployment SPI
module. This is technically a breaking change, because they are also moved
to a different package, but these build items don't seem to be used anywhere
outside of Quarkus, so it should be safe.