Redis Client: add runtime SPI and deployment SPI modules
Some constants from the Redis Client runtime module are moved
to the runtime SPI module, so that they can be used from other modules
that cannot directly depend on the Redis Client extension.

One build item, `RequestedRedisClientBuildItem`, is moved from the Redis Client
deployment module to the deployment SPI module. This is technically a breaking
change, because it is also moved to a different package, but this build item
doesn't seem to be used anywhere outside of Quarkus, so it should be safe.

Further, one new build item, `RedisClientBuildItem`, is added. It provides
runtime access to the Redis clients (in the Mutiny variant) without having
to perform a CDI lookup.