Get rid of {doc-guides} attribute in documentation
It has been mistakenly used in the newest guide that was added as shown in
https://github.com/quarkusio/quarkus/pull/36474 where we had:
- one invalid link because of a missing /
- one invalid link because an incorrect attribute was used

We see it's risky and it has no more value since
we introduced the downstream archive as we automatically transform the
links there to point to either a local one or one that is upstream.

IIRC, Michelle already removed some and I don't think it's recommended
nowadays.

Note: I haven't handled the getting-started-devservices.adoc here on
purpose as Leonor is working on it. Will handle it if needed once
Leonor's PR is in.