Redis: use pre-existing command instances instead of using Command.create()
This is more idiomatic and prevents troubles with Redis cluster where
the Vert.x Redis client needs to know the key affected by the command
in order to route the command to the correct cluster node. That works
correctly with pre-existing `Command` instances and will _not_ work
correctly with `Command.create()` until the Vert.x Redis client itself
is fixed.