Redis: normalize results of `Command.create()`
When the `Command.create()` method is used to create a `Command` instance
for a known command (not recommended, but also not forbidden), it returns
a generic instance that doesn't know anything about where the keys are in
the command. Such generic instance is unusable with Redis cluster, because
the target node will be selected randomly, not based on the key, and there's
high chance such command will result in the `MOVED` redirect.

With this commit, we normalize all results of `Command.create()` so that
a pre-existing static instance for known commands is used. This is key-aware
and works with Redis cluster.

This commit adapts a fix submitted to Vert.x Redis client. When we upgrade
to a fixed Vert.x Redis client, this commit should be reverted.