Fix restore happening in the same thread
When the restore is performed from the same thread, creating Vertx
HttpServer results in it being registered under the same
io.vertx.core.Context (cached in a thread-local). That way the Netty
eventloop is also the same and the requests are not balanced across
threads - causing performance regression as request handling becomes
effectively single-threaded.