ArC: fix in-application build compatible extensions
In case the _application_ contains a build compatible extension that
registers a synthetic bean for a _non-application_ class, or a synthetic
observer declared "as if" in a _non-application_ class, the generated
classes must be forced to also be application classes. If the generated
classes were non-application classes (as they were prior to this commit),
they would not see the creator/disposer/observer classes in Quarkus
dev mode and test mode (where application classes are in a different
class loader than non-application classes).

Before this commit, it has already been possible to force ArC to generate
synthetic bean classes as application classes, but this was not possible
for synthetic observers. This commit adds that, too.