ArC: fix ambiguity resolution in case of producers with priority
A producer may have priority as defined by CDI 4.1 (or obtained from
a stereotype, as permitted but not really specified in CDI 4.0),
but does not necessarily have to be an alternative. During ambiguity
resolution, if such non-alternative producer with priority is declared
on an alternative bean, we must first consider the priority declared
on the producer, even if the producer is not itself an alternative.

To do that, this commit changes the ambiguity resolution implementation
to use `BeanInfo.getPririty()` instead of `getAlternativePriority()`.
The difference is that `getAlternativePriority()` returns `null` when
the bean itself is not an alternative. Given that the method was
previously only used during ambiguity resolution, and only after
non-alternative beans (and producers on non-alternative beans)
were eliminated, this is correct.

In fact, this commit also marks `BeanInfo.getAlternativePriority()`
as deprecated for removal. For all practical purposes, `getPriority()`
is a better choice; it is usually meaningless to consider a bean
priority only when the bean is an alternative, and pretend that
the bean has no priority when it is not an alternative.