ArC: prevent duplicate bean discovery & detect duplicate bean identifiers
A bean class may occur in the bean archive index multiple times, because
the index is in fact a composite of multiple indices and there's nothing
preventing duplicity in the composition.

Fortunately, all the additional customizations (such as default scope or
unremovability from `AdditionalBeanBuildItem`) are passed around outside
of the bean archive index (for example, as annotation transformations or
as `BeanProcessor` configuration), so they are not lost.