Netty requires more classes to be runtime initialized
In https://github.com/quarkusio/quarkus/pull/37347 we moved
`PlatformDependent` and `PlatformDependent0` classes to
run-time-initialization as they are platform-dependent and thus need to
be initialized on the target platform. It looks like there are more
classes like these that need to be runtime initialized as they
transitively rely on platform dependent values.

Closes https://github.com/Karm/mandrel-integration-tests/issues/236

Supersedes https://github.com/quarkusio/quarkus/pull/37628