Avoid a warning for tests in interactive mode
The CliHelpTest#testCommandHelp has become extremely flaky lately
because for one run, a warning saying JBang is not installable is there
and for the other it's not.

This is problematic and I really don't see the value of adding a warning
about JBang that is specifically only displayed for tests.

Also fixed a bunch of typos and made the message more neutral (let's
avoid using exclamation marks for messages and exceptions).