New Route SPI
Unlike the current one, this SPI allows declaring routes without having
to depend on Quarkus Vert.x HTTP. Thus, if an extension declares a route
with this SPI, and Quarkus Vert.x HTTP is not present, the route is ignored.
However, if Quarkus Vert.x HTTP is present, the route is registered.