Use UpdateDependencyVersionOperation first to update Quarkus version
It is important to try that first as that recipe updates the OpenRewrite
model, and thus the BOM, for the following recipes.
This is important for the recipe syncing the Hibernate ORM modelgen
annotation processor with the version coming from the BOM as we want to
sync it with the updated version.

(cherry picked from commit f2fda625c9a0cfb52bca816835aa5a79f24305b7)