Fix request hanging condition
This condition could happen when a TLS secured connection had a lot  of data written to it
very fast from an event loop thread.

The problem was that there was a race condition
on the registration of the drain handler,
which then resulted in the pending data never
being written out.
The fix is to always register the drain handler

Resolves: #37807