Add a tool to check cross references
Unfortunately, it doesn't support additional ids added just after the
title as in:
[id1]
== [id2] [id3] Title

so it will force people to use the canonical id but I think it's
acceptable in our reference doc to enforce the canonical id.

Also move the link transformation to downstream documentation only as
this alters the ability of Asciidoctor to check the internal links.