Revert "Fix static JPA metamodel generated for Panache classes"
Removes workaround for HHH-17683 in Panache.

We need to keep specifying the version for Kotlin extensions due to
kotlin-maven-plugin not taking dependency management into account;
see
https://github.com/quarkusio/quarkus/issues/37477#issuecomment-1923662964.
However, we now use the same version as the one we expect people to use
in applications (6.4.3), so it's no longer so much of a hack.

This reverts commit f4b489bc23e121daf53db42a923c8acf5c344889.
This reverts commit e745fe32569010f7a0a5e1eedf0b4dcdffa881e6.