Allow TLS certificate reloading for the HTTP server
Key store, trust store and certificate files can be reloaded periodically.
The period is configured using the `quarkus.http.ssl.certificate.reload-period` property.

The files are reloaded from the same location as they were initially loaded from.
If there is no content change, the reloading is a no-op.
It the reloading fails, the server will continue to use the previous certificates.