Updated MongoDB/Panache parser to latest ORM parser
This way, we keep it up to date, because the previous version was a
pre-release of ORM 5-6.
This required some visitor changes, and in particular there's a
precedence change between 'and/or' expresions who now require paren
grouping.
Also, literals are not implemented correctly: only strings are supported
but this was already the case: tests do not test other literals.