ArC: fix creation of synthetic beans
If a synthetic bean creation function is backed by a `Supplier`, `Function`,
`RuntimeValue` or a runtime proxy, that creation function is stored into
a `Map` during application startup and later retrieved when needed.

The key under which it is stored into the `Map` is wrong, as it only contains
the name of the implementation class and a hash of bean types and qualifiers.
When there are multiple synthetic beans with the same implementation class,
types and qualifiers, only one creation function is used for all of them.

This commit fixes that by including the synthetic bean identifier in the hash,
making creation function key unique for each synthetic bean.

(cherry picked from commit b0b898aac5a918d5cc2eade0e7a460e4c5ba7a04)