Fix misleading error message when REST Client interface has been indexed
When a REST Client interface is part of a library or the runtime part
of an extension, it's not uncommon to forget to index the module.
When that happens, we now provide a proper error message.

(cherry picked from commit 573e595bfc40e1460ef394d2a9aae9259f3ac04b)