Make sure the scripts listing dependencies are locale independent
Katia reported some issues when trying to update the listed dependencies
with the scripts because she was using a locale with a different order
than the one used to generate the current content.
Let's use the C locale as it's the most basic one and should be around
on all boxes.