CP: set up a temporary no-context/no-executor ContextManagerProvider during boot
Because some extensions (only known one is spring-cloud-config-client)
use Vert.x via Mutiny before runtime init is done. Mutiny and CP are
both not set up yet, due to missing executor, but I'm starting CP with
zero contexts and an executor which throws with an explicit error
message, which should be enough to boot until runtime is properly set
up.