Rewrite footnotes for downstream documentation
The modern version of footnotes is not supported in the downstream
tooling and we can't use the older version in our doc as it triggers
warnings.
Thus we rewrite the new format to the old format specifically for
downstream doc.

(cherry picked from commit 82f7b20709255ea7e08aa8237eee0aba41903a91)