Fix resource registration for native compilation
NativeImageResourceBuildItem and ServiceProviderBuildItem contain a path
of the resource we should include and not a pattern or a glob. As a
result, `Pattern.quote` is the right method to use in order to produce a
pattern that would match the path.

The patch also remove the wrong addition of the paths "as is" to the
includes json array.

Fix up of b7f49dd9d952a46fde3f0abaabcc08b1a2e5f067

(cherry picked from commit 2585bb0c5a8e8e4e758d4d6e49d98b13380267f6)