Refactor quarkus.tls.trust-all to use new TLS registry
This commit replaces the build-time property 'quarkus.tls.trust-all' with a new TLS registry. The previous implementation was potentially problematic, as it could lead to security issues in a production environment.

Please note, this is a breaking change for extensions that use the now-deleted `TlsConfig.java`. However, from a user perspective, the change is seamless. The 'quarkus.tls.trust-all' configuration property has simply transitioned from a build-time to a runtime configuration.