Significantly reduce the number of interfaces registered for config
We only need to register for reflection config interfaces that are
constrained. We cannot be more fine grained at the moment but
theoretically we could only register the methods required to go from the
root of the config mapping to the (potentially nested) constrained
interface.
Also make sure config implementors can be properly handled by including
them in the indexes considered by HV.