Allow Snappy to be loaded from a shared classloader
This commit introduces a solution for loading the Snappy native library across multiple test profiles. Due to the constraint that native libraries can only be loaded from a single classloader, a shared classloader is now utilized for loading Snappy in test mode.

Please note, this feature is exclusively applicable when running tests.

Fixes: https://github.com/quarkusio/quarkus/issues/39767