Add a temporary config property to allow multiple resources
The plan is to remove this property in 3.11 but we need to provide a way
for people to update to latest security fixes without having to
significantly adjust their applications.

In 3.11, we need to document this breaking change properly and also
provide guidance on how to fix the most common issues that could be
encountered due to this breaking change.

(cherry picked from commit aa7d2ee19afe13b01cabf5e7bbbf47d243d88a3f)