Make sure we transmit the actual debug port to next dev mode run
When starting dev mode, the debug port might not be free and we might
use a random one.
When restarting dev mode (for instance when you adjust the pom.xml), we
should try to reuse this port instead of defaulting to the initially non
free one.

Also the port being free is tested when we build `DevModeRunner`, so we
need to make sure we stop the old runner before build()ing the new
instance (and not just before we start the new one).

Fixes #40848

(cherry picked from commit 2233494adc057c6b74140209677a72cb23a321cf)