Remove image metrics integration tests
As part of https://github.com/quarkusio/quarkus/issues/40076 is now
implemented and we are gathering the necessary stats in a database there
is no longer the need for these tests to exist.

Note that automated alerting in case of regressions is still not setup
but we now have a way to go back in time and see which (few) patches
resulted in the regression, even if we don't spot it immediately.