Simplify code in RestAssuredURLManager
The code still works whether RestAssured is on the classpath
or not, but it now contains direct invocations of RestAssured
code instead of reflection.
It goes without saying that this make the class a lot more
maintainable

At some point in the future, we probably want to have some
sort of SPI, but for now, this change is much better than the
current state.