Update to gRPC 1.65.0 and Netty 4.1.111
This commit updates the gRPC version to 1.65.0 and enables the switch to Netty 4.1.111. Additionally, it updates several gRPC-related dependencies:
- jprotoc to 1.2.2
- protoc to 3.25.3
- protobuf-java to 3.25.3

Disable Stork "least response time" IT when using the gRPC Java client. This is because in gRPC 1.65.0, the load balancer is only called once if the calls are queued. The Vert.x-based client is still behaving normally. When using gRPC Java client, it's a change in behavior, but the load balancer will still be called for non-concurrent calls.