Clear com.sun.naming.internal.ResourceManager#propertiesCache
This is only effective if the class is made accessible, which we make in
our code.
I noticed this a few times in my heap dumps and I noticed this was also
handled in some other frameworks (until the switch to Java 9 that made
the --add-opens mandatory).
I think dealing with it in our codebase is a good idea from a
sustainability point of view.