use recognized status
"draft" was used as status which is not a recognized stated.

Updating to "experimental" as that's what we document in https://quarkus.io/guides/writing-extensions and similar.

But I don't know if that is the true state.

(cherry picked from commit 393b8800a2af8b87ad8d536871d2a929d0350968)