Make config annotations kept at runtime
This should allow us to resolve fully the doc model even if we depend on
classes or config groups that are outside of the scope of the module.
The cost should be negligible and it should simplify things a lot.

All the @ConfigMapping-related annotations in SmallRye Config already
have a runtime retention so it makes sense to be consitent.

Note that we will still need to centralize the Javadoc as obviously the
Javadoc is source only.