Improve how ConfigSections/ConfigRoots are merged
ConfigSections have to be merged inside a given ConfigRoot. Typically,
HibernateORMConfig has sections that are both in a simple element and in
an unnamed map thus leading to the same section path.

As for ConfigRoots, let's not merge them at all at resolution stage,
given we already do some merging in the Maven plugin,
let's do the merging at the end.

Also drops some code related to having ConfigGroups in the model as we
don't need them.