Allow not detecting the extension
We actually have some level of support for building an extension with
Gradle so we need to make it work.
However, at the moment, it's close to impossible to determine the
extension we are in in a Gradle extension.
So we allow not detecting the extension and disable the config doc
generation in this case.