Do not keep the whole Manifest in memory
Some manifests are quite large, for instance the one for commons-codec
is 21 KB. Given most of the information in the Manifest are of no
interest to us, let's trim the Manifest to only the information we need.

Also we still have the problem of us keeping the jars open multiple
times in test and dev mode so it amplifies this issue.

In some cases, I ended up with 1700+ jars opened so the Manifests add up
quite quickly.