Use the root path for OpenContainerPathTree#getRoots()
The container path is the path of the container (could be a directory or
an archive and shouldn't be used for browsing.
The root path is the path of the root. In the case of an archive, it's
actually a ZipPath, which can be used to browse the archive.

Fixes #42571

(cherry picked from commit 6dd61509ce9925d701fe1589a2304ee07a848400)