Start Testcontainers containers with the Platform CL as TCCL
Testcontainers start several long lived threads (Ryuk reaper, ducttape),
and if the TCCL is one of the QuarkusClassLoader, we end up leaking the
class loaders.
Using the Platform CL makes sure we don't leak any QuarkusClassLoader
via these threads.

The observability containers have not been adjusted as they are a bit of
a weird beast and not worth having a look in the scope of this
experiment.