Include the run attempt in the build-reports file name
When we have multiple attempts of a given workflow run (for instance,
when we manually rerun it), only the first build-reports artifacts
are saved as artifacts are immutable.

Thus we keep generating the report based on the artifacts of the first
run, which is incorrect and confusing.

This should be backported to all supported branches as handling the
transition is not exactly easy.