Remove a hack introduced for Liquibase that is no longer needed.
With the new index, we don't need that anymore. I tested that the test
introduced in this commit is still working fine
(LiquibaseExtensionMigrateAtStartDirectoryChangeLogTest).

Note that I fixed it as I forgot to consider the less priority elements
there and I went for fixing it and then was wondering why this would be
even useful.

I also introduced a shortcut for empty resource.