OIDC: Prevent concurrent tenant creations
Refactors `TenantConfigBean` internally to both prevent concurrent creation of the same OIDC tenant and optimizes the implementation to use non-volatile fields.

Tenant contexts are managed via the map-container (`TenantsMap`) and map-value type (`TenantHolder`), both are optimized for the "99.99% case" that the maps of static and dynamic tenants are "fully populated" and all (active) tenants have a "ready" context.

Retrieving a tenant hits the "fast path" in the vast majority of all invocations, accessing only non-volatile fields and a non-concurrent hash map.

In case a tenant is not yet present (new dynamic tenant ID or late-initialized static default tenant), the implementations switches to the "slow path" and creates a new hash map with the new tenant and updates the non-volatile field referencing the tenants-map. Other threads that might not yet "see" the updated tenants-map field would also enter the slow path - but once the slow path has been executed, the update to the tenants-map field will be "visible".

The `TenantHolder` uses a mixture of non-volatile (the "99.99% case" is accessing a ready context) and volatile fields. The latter are used to make changes immediately "visible" to other threads - the order of writes is important here. Unlike `TenantsMap`, the `TenantHolder` class does not use/need a lock object or monitor.