Update rest-data-panache docs to clarify experimental status
The Panache extension for generating CRUD resources is split into 3
packages, the most-used one of which is stable. But readers of the guide
are greeted with a warning claiming the feature is experimental, which
is driving teams away from using it. This patch removes the warning and
clarifies in the compatibility table the status of each package, as per
https://quarkus.io/extensions/?search-regex=panache&status=stable